+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activitys.MainActivity">
+        <activity android:name=".activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -23,17 +23,17 @@
         </activity>
 
         <service
-            android:name=".services.FloatingService"
+            android:name=".service.FloatingService"
             android:enabled="true"
             android:exported="true">
         </service>
 
-        <receiver android:name=".receiver.NetworkStateReceiver">
+        <receiver android:name=".receivers.NetworkStateReceiver">
             <intent-filter>
                 <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
             </intent-filter>
         </receiver>
-        <receiver android:name=".receiver.ApkInstallationListener">
+        <receiver android:name=".receivers.ApkInstallationListener">
             <intent-filter>
                 <action android:name="android.intent.action.PACKAGE_ADDED" />
 
@@ -42,7 +42,7 @@
         </receiver>
 
         <service
-            android:name=".services.MyAccessibilityService"
+            android:name=".service.MyAccessibilityService"
             android:enabled="true"
             android:exported="true"
             android:label="@string/smart_service"
