+++ app/src/main/java/com/example/liangweiwu/downloadmanager/threads/DownloadTaskPoolThread.java
@@ -1,10 +1,10 @@
 package com.example.liangweiwu.downloadmanager.threads;
 
-import com.example.liangweiwu.downloadmanager.activitys.events.MainUiEvent;
+import com.example.liangweiwu.downloadmanager.activity.events.MainUiEvent;
 import com.example.liangweiwu.downloadmanager.model.ApkInformation;
 import com.example.liangweiwu.downloadmanager.model.DownloadTaskController;
-import com.example.liangweiwu.downloadmanager.utils.FileUtils;
-import com.example.liangweiwu.downloadmanager.utils.GameInformationUtils;
+import com.example.liangweiwu.downloadmanager.util.FileUtils;
+import com.example.liangweiwu.downloadmanager.util.GameInformationUtils;
 
 import java.util.ArrayList;
 import java.util.Iterator;
